Implement registration functionality
In the Core application, add urls.py and include the urls from
the Core application in the todolist/urls.py file. Add a
serializers.py file to the Core application. Define a
ModelSerializer for registration (check the password strength
using Django's built-in password validation, accept 'password'
and 'password_repeat' fields and check if they match; do not
store the password in plain text, but save its hash).